Added tests for inout controller, rotating controller, multi learning controller
InOutController has been changed to fit in to the new test. Now get_output method
needs a tuple for all incoming values. The new test contains a check whether
get_output and get_control is implemented. Only set_control is not checked
because checking it is almost impossible without a helper variable.
Fixed unwanted shared variable behaviour by defining the variable inside init.

RotatingController is a class split from MultiLearningController (a.k.a.
LearningMultiController but this name is not correct anymore, will think of a
different name later on). This class contains the methods how the controllers
will be added and selected.
The BaseTest will test all the methods' inheritance of RotatingController.

MultiLearningController need more tests for testing the reflecting method.
It already inherits two BaseTests.

Commits will be shorter but not the messages