Refactoring all gyms and rotating learning controller
All gyms have a shared coding style now. Starting with control tuples of each
pid and node controller. This gets added to a RotatingLearningController
object.
In every EnvironmentController the output gets combined. These steps step
will get a overhaul in the future because this is tedious. The rewards are
processed for every pid or node.

Refactored LearningMultiController to RotatingLearningController as this
the class rotates LearningControllers instead of learning MultiControllers.