Refactoring environment runner
Changed EnvironmentRunner to EnvironmentManager because the class is doing too many for just being a runner.
Changed learner to agent. Agent is doing things and will reflect and explore new options.
Changed controller to worker. The worker will get the action and reward from the observation.
Changed EnvironmentController to EnvironmentWorker. The worker has not too many difficult tasks but is still important.

Added method run(). This reduce the redudant code in the environment code.
Changed run_once() to use a while loop because not every environment is the same.

Tkinter framework option control

Option panel takes values of an info dictionary and put them in the interface. Refresh button takes the values from the info dictionary to the interface. Update button takes the values from the interface to the info dictionary.
Possible info the option panel can display: Texts, numbers, on/off.

Added custom Spinbox class.
Added custom spinbox to Frame class.
Removed can_update and can_restore functions, might remove the variables in the future. It is not clear how the current state should be saved and how to restore the previous state at start up.